refactor, add tests
- refactor so that necessary functions can be mocked

- add tests to ensure code runs to completion

- add dummy data to data folder

- add test for reading csv files in directory
