+++ src/PatientX/models/MistralRepresentation.py
